+++ src/pages/Main/Main.js
@@ -1,6 +1,5 @@
-import React, { useEffect, useState } from 'react';
-import { Link } from 'react-router-dom';
-
+import React from 'react';
+import Products from './Products';
 import './Main.scss';
 
 const Main = () => {
@@ -47,36 +46,4 @@ const Main = () => {
     </>
   );
 };
-
-const Products = () => {
-  const [productData, setProductData] = useState([]);
-  useEffect(() => {
-    fetch('data/productData.json')
-      .then(res => res.json())
-      .then(data => {
-        setProductData(data);
-      });
-  });
-  return (
-    <>
-      {productData.map(({ id, img, title, hashtxt1, hashtxt2 }) => {
-        return (
-          <div className="productContainer" key={id}>
-            <Link to="#" className="productItemLink">
-              <div className="productItemThumb">
-                <img src={img} alt="thumb1" />
-              </div>
-            </Link>
-            <div className="productItemHashWrap">
-              <span className="productItemHash">{hashtxt1}</span>
-              <span className="productItemHash">{hashtxt2}</span>
-            </div>
-            <div className="productItemName">{title}</div>
-          </div>
-        );
-      })}
-    </>
-  );
-};
-
 export default Main;
